+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - add and remove tags to JPEG/GIF/PNG/TIFF images
 - French translation
+- restored support for Android KitKat (without Google Maps)
 
 ## [v1.5.6] - 2021-11-12